requirements.txt
beautifulsoup4==4.11.1
cfscrape==2.1.1
Flask==1.1.2
Pillow==9.2.0
SQLAlchemy==1.4.31
